backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
requests==2.32.3
python-dateutil==2.9.0
pydantic==2.9.2